Add new fighter count override entity

+++ dh-adapted/src/handler_json/data/primitives.rs
@@ -11,4 +11,4 @@ pub(in crate::handler_json) type CItemCatId = rc::ad::AItemCatId;
 pub(in crate::handler_json) type CAttrVal = rc::ad::AAttrVal;
 pub(in crate::handler_json) type CCount = rc::ad::ACount;
 pub(in crate::handler_json) type CSlotIndex = rc::ad::ASlotIndex;
-pub(in crate::handler_json) type CSkillLevel = i8;
+pub(in crate::handler_json) type CSkillLevel = i32;
