+++ http/src/shared/primitives.rs
@@ -1,3 +1,3 @@
 pub(crate) type HMutaRoll = f64;
-pub(crate) type HSkillLevel = i8;
+pub(crate) type HSkillLevel = i32;
 pub(crate) type HFitSecStatus = f64;
